backend/requirements.txt
# --- Core Backend ---
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
requests==2.32.3

# --- Database ---
SQLAlchemy==2.0.25

# --- AI & Vector Search ---
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.5
faiss-cpu==1.7.4
sentence-transformers==3.0.1

# --- LLMs ---
langchain-groq==0.1.3

# --- File Handling & Parsing ---
pdfplumber==0.11.0
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==10.3.0

# --- Cloud Integration ---
supabase==2.5.1

# --- CORS & Middleware ---
starlette==0.37.2

# --- Optional Utilities ---
python-multipart==0.0.9
